requirements.txt
# Django Core
Django>=5.0.4
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
mysqlclient>=2.2.0
psycopg2-binary>=2.9.9

# Async Tasks
celery>=5.3.0
redis>=5.0.0

# AI/LLM
pyautogen>=0.2.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0

# File Processing
Pillow>=10.0.0

# WebSocket (optional)
channels>=4.0.0
channels-redis>=4.1.0

# Development
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0
